Added vendors to Dungeon Spiral!
Vendors appear as a blue @ randomly in the depths of the dungeon.
Access their wares by pressing 's' while atop their locations.

Added a currency (cashola) to make transactions possible. Currently, all items in the game use a default price of 1, making the vendors essentially a simple barter system for the time being, though the interface to adjust the pricing of items at will is in place and will not be difficult to implement. It will probably help to make merchants charge more than they pay as well.

Secondary update:
Introduced the wand of blood attack wand. It leeches life from nearby enemies and heals the player for damage done.